kong_secure_api_platform/microservice/app/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
PyJWT==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6
